Dont analyze any files when cc config include_paths is empty list
When a user runs `codeclimate analyze <path>`, workspace paths to analyze get passed
into `config.json` as a list.

The previous implementation made the mistake of expecting an empty list to be truthy,
and consequently thought the `include_paths` key was missing if include_paths == [].

As a result, it analyzed all files in directory instead of none.

This change fixes that behavior by checking the type of object that's returned
by config.get("include_paths").